requirements.txt
pandas 
numpy 
matplotlib 
seaborn 
langchain 
sentence-transformers 
faiss-cpu
transformers
torch